+++ Core/TI59Machine.cpp
@@ -241,9 +241,9 @@ void TI59Machine::clearDebugOverlay() {
 }
 
 bool TI59Machine::runDebugOverlay(uint16_t startAddr, uint32_t maxSteps,
-                                  uint32_t* outSteps, bool* outSawHold) {
+                                  uint32_t* outSteps) {
     std::lock_guard<std::mutex> lock(m_keyMutex);
-    return m_cpu.runDebugInjectedProgram(startAddr, maxSteps, outSteps, outSawHold);
+    return m_cpu.runDebugInjectedProgram(startAddr, maxSteps, outSteps);
 }
 
 uint32_t TI59Machine::drainCpuFrames(CpuFrame* out, uint32_t max, uint32_t* outLost) {